requirement.txt
anyio==4.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
exceptiongroup==1.1.3
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
loguru==0.7.0
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
python-dotenv==1.0.0
python-telegram-bot==12.0.0b1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tornado==6.3.3
urllib3==1.26.16